Final merge from branches/merge/ticket_29_emote3_adc
git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@4312 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/merge/tstp/include/chronometer.h
@@ -64,8 +64,7 @@ public:
     void lap() { if(_start != 0) _stop = Alarm::_elapsed; }
     void stop() { lap(); }
 
-    // The cast provides resolution for intermediate calculations
-    Microsecond read() { return static_cast<unsigned long long>(ticks()) * 1000000 / frequency(); }
+    Microsecond read() { return ticks() * 1000000 / frequency(); }
 
 private:
     Time_Stamp ticks() {
